Collecting parquet annotations and storing them in the map
Implemented SchemaUtil deepCopy with parquet annotations enrichment

Fix errors

Added more tests

Improve error message (#611)

Make implicit field from EnumType low priority (#613)

Fix build
